backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Document parsing (requires poppler on Windows - see README)
unstructured[pdf,docx,pptx]>=0.12.0

# Vector store
chromadb>=0.4.0
openai>=1.0.0

# MS Graph (for live mode)
msal>=1.26.0
msgraph-sdk>=1.0.0

# Agent framework
camel-ai>=0.2.0

# WebSocket
websockets>=12.0

# Async support
aiofiles>=23.0.0
